# Optimize headerfile write during endpoint regeneration