… pkg/endpoint/restore_test.go