This PR optimizes the headerfile writing portion of endpoint regeneration, to reduce the amount of data that is written to disk when the headerfile must be written, and to reduce the frequency of disk writes (which results in a corresponding improvement in regeneration time).

More specifically:
* Don't serialize endpoint status. This field is not useful after endpoint restore, so skip writing it.
* Remove deprecated options format. Cilium 1.1 or earlier had another format for these, but we don't need to serialize this any more.
* Only rewrite headerfile when the endpoint changes. I'll include the commit log for better review towards the end of this PR description.

Using the identity churn test in #8063, I set up 15K identities, then inserted a further 10K which are selected by an endpoint, over the course of about ten minutes. The screenshots below include 90th and 99th percentile regeneration times, where I would expect that the 99th relates to the endpoint which has a policy allowing connectivity to the identities being added, and the 90th relates to endpoints which do not. Around 8-10 endpoints are present during the test.

There's an anomaly at the beginning of the test which we're yet to narrow down, but note two things: Firstly, towards the end of the test the `prepareBuild` label reports 31ms of time spent per endpoint build in the 99th percentile (average time over 2 minutes of build, estimated 200+ regenerations). Secondly, the overall `buildDuration` is 45ms. Looking also at the 90th percentile, the average `buildDuration` is close to 15ms for much of the test.

![master_20190610_add_10k_selected](https://user-images.githubusercontent.com/1243336/59235165-da026c00-8ba4-11e9-8081-ba3434efa98a.png)

Compare with this PR, where `prepareBuild` takes around 8.5ms and the overall `buildDuration` has cut to 31ms for the 99th percentile. In the 90th percentile, average `buildDuration` drops to around 8ms.

![pr_8268_20190610_add_10k_selected](https://user-images.githubusercontent.com/1243336/59235377-b2f86a00-8ba5-11e9-8d47-2abc2683d290.png)

Fixes: #8144 

---

Last commit message: 

    The endpoint headerfile is used for two purposes:
    (1) to provide static data to compile/rewrite/load the datapath for the
        endpoint; and
    (2) to store information about the endpoint across restart/upgrade.

    This commit relies upon loader.EndpointHash(e) to determine whether any
    datapath variables have changed, and hence whether to rewrite the
    endpoint headerfile. While this strictly speaking only takes into
    account the (1) purpose above, in practice this covers almost every
    field that may change after the initial regeneration (where the
    headerfile would always be written):

| Field                 | Part of EndpointHash()? | Changes at runtime? |
| --- | --- | --- |
| ID                    | No                      | No                  |
| ContainerName         | No                      | No                  |
| dockerID              | No                      | No                  |
| DockerNetworkID       | No                      | No                  |
| DockerEndpointID      | No                      | No                  |
| DatapathMapID         | Yes                     | Yes                 |
| IfName                | No                      | No                  |
| IfIndex               | Yes                     | No                  |
| OpLabels              | No                      | Yes(1)              |
| LXCMAC                | Yes                     | No                  |
| IPv6                  | Yes                     | No                  |
| IPv4                  | Yes                     | No                  |
| NodeMAC               | Yes                     | No                  |
| SecLabel              | Yes                     | Yes                 |
| Options               | Yes                     | Yes                 |
| DNSHistory            | No                      | Yes(2)              |
| K8sPodName            | No                      | No                  |
| K8sNamespace          | Yes (Part of SecLabel)  | No                  |
| DatapathConfiguration | Yes                     | Yes                 |

    Any field that does not change during runtime only needs to be written
    to the filesystem once, and can be ignored as part of the calculation
    for whether to rewrite the header file.

    Any field that is part of the endpoint hash will trigger this logic to
    ensure that the correct new value is written to the filesystem.

    This leaves only a couple of interesting cases:

    (1) Operational Labels: These are not typically changed during runtime,
        but even if they are, given the container runtime ID we should be
        able to query the correct set of labels so it's not the end of the
        world to lose this information. These labels have been determined
        not to be security relevant based on the fact they're in this field.

    (2) DNS History: endpoint.SyncEndpointHeaderfile() contains a controller
        which is responsible for periodically synchronising the header file
        on the filesystem with the latest DNS history for the endpoint, so
        it is safe to ignore this field during regeneration cases.

    Given that these cases are otherwise handled, rely upon the EndpointHash
    to determine whether to write the header file to the filesystem.

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/cilium/cilium/8268)
<!-- Reviewable:end -->